reorganize code for sending near transactions
