+++ node/src/indexer/tx_signer.rs
@@ -1,3 +1,5 @@
+use crate::config::RespondConfigFile;
+use crate::indexer::types::ChainSendTransactionRequest;
 use near_crypto::{InMemorySigner, PublicKey, SecretKey, Signer};
 use near_indexer::near_primitives::account::AccessKey;
 use near_indexer_primitives::near_primitives::transaction::{
@@ -5,6 +7,7 @@ use near_indexer_primitives::near_primitives::transaction::{
 };
 use near_indexer_primitives::types::AccountId;
 use near_indexer_primitives::CryptoHash;
+use std::sync::Arc;
 use std::sync::Mutex;
 
 pub(crate) struct TransactionSigner {
@@ -64,5 +67,65 @@ impl TransactionSigner {
 
     pub(crate) fn public_key(&self) -> PublicKey {
         self.signer.public_key()
+    }
+}
+
+pub(crate) struct TransactionSigners {
+    /// Signers that we cycle through for responding to signature requests.
+    /// These can correspond to arbitrary near accounts.
+    respond_signers: Vec<Arc<TransactionSigner>>,
+    /// Signer we use for signing vote_pk, vote_reshared, etc., which must
+    /// correspond to the account that this node runs under.
+    owner_signer: Arc<TransactionSigner>,
+    /// next respond signer to use.
+    next_id: usize,
+}
+
+impl TransactionSigners {
+    pub fn new(
+        respond_config: RespondConfigFile,
+        owner_account_id: AccountId,
+        owner_secret_key: SecretKey,
+    ) -> anyhow::Result<Self> {
+        let respond_signers = respond_config
+            .access_keys
+            .iter()
+            .map(|key| {
+                Arc::new(TransactionSigner::from_key(
+                    respond_config.account_id.clone(),
+                    key.clone(),
+                ))
+            })
+            .collect::<Vec<_>>();
+        let owner_signer = Arc::new(TransactionSigner::from_key(
+            owner_account_id,
+            owner_secret_key,
+        ));
+        anyhow::ensure!(
+            !respond_signers.is_empty(),
+            "At least one responding access key must be provided",
+        );
+        Ok(TransactionSigners {
+            respond_signers,
+            owner_signer,
+            next_id: 0,
+        })
+    }
+
+    fn next_respond_signer(&mut self) -> Arc<TransactionSigner> {
+        let signer = self.respond_signers[self.next_id].clone();
+        self.next_id = (self.next_id + 1) % self.respond_signers.len();
+        signer
+    }
+
+    fn owner_signer(&self) -> Arc<TransactionSigner> {
+        self.owner_signer.clone()
+    }
+
+    pub fn signer_for(&mut self, req: &ChainSendTransactionRequest) -> Arc<TransactionSigner> {
+        match req {
+            ChainSendTransactionRequest::Respond(_) => self.next_respond_signer(),
+            _ => self.owner_signer(),
+        }
     }
 }
